nds/pyproject.toml
[project]
name = "nds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "grpcio-tools>=1.67.1",
    "grpcio>=1.67.1",
]
